pyproject.toml
[tool.poetry]
name = "-"
version = "0.1.0"
description = ""
authors = ["Isis e Celly"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
databases = "^0.9.0"
mysql-connector-python = "^8.3.0"
aiomysql = "^0.2.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
